requirements.txt
pyodbc
streamlit
pandas
python-dotenv
psycopg2-binary
groq
